+++ src/main/java/io/apiman/cli/command/declarative/command/AbstractApplyCommand.java
@@ -24,22 +24,14 @@ import io.apiman.cli.command.declarative.DeclarativeUtil;
 import io.apiman.cli.command.declarative.model.BaseDeclaration;
 import io.apiman.cli.exception.CommandException;
 import io.apiman.cli.services.WaitService;
-import io.apiman.cli.util.BeanUtil;
 import io.apiman.cli.util.MappingUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 import java.util.stream.Collectors;
-
-import static java.util.Optional.ofNullable;
 
 /**
  * Applies an API environment declaration.
@@ -88,29 +80,11 @@ public abstract class AbstractApplyCommand extends AbstractFinalCommand {
 
     /**
      * Load and then apply the Declaration.
+     *
      * @param declarationFile
      */
     private BaseDeclaration loadDeclaration(Path declarationFile) {
-        final Map<String, String> parsedProperties = BeanUtil.parseReplacements(properties);
-
-        // check for properties file
-        ofNullable(propertiesFiles).ifPresent(propertiesFiles -> propertiesFiles.forEach(propertiesFile -> {
-            LOGGER.trace("Loading properties file: {}", propertiesFile);
-
-            final Properties fileProperties = new Properties();
-            try (final InputStream propertiesIn = Files.newInputStream(propertiesFile, StandardOpenOption.READ)) {
-                if (propertiesFile.toAbsolutePath().toString().toLowerCase().endsWith(".xml")) {
-                    fileProperties.loadFromXML(propertiesIn);
-                } else {
-                    fileProperties.load(propertiesIn);
-                }
-            } catch (IOException e) {
-                throw new CommandException(String.format("Error loading properties file: %s", propertiesFile), e);
-            }
-
-            fileProperties.forEach((key, value) -> parsedProperties.put((String) key, (String) value));
-        }));
-
+        final Map<String, String> parsedProperties = DeclarativeUtil.parseProperties(properties, propertiesFiles);
         final BaseDeclaration declaration = loadDeclaration(declarationFile, parsedProperties);
         LOGGER.info("Loaded declaration: {}", declarationFile);
         LOGGER.debug("Declaration loaded: {}", () -> MappingUtil.safeWriteValueAsJson(declaration));
@@ -124,13 +98,13 @@ public abstract class AbstractApplyCommand extends AbstractFinalCommand {
     protected abstract void applyDeclarations(List<BaseDeclaration> declaration);
 
     protected BaseDeclaration loadDeclaration(Path declarationFile, Map<String, String> parsedProperties) {
-      // parse declaration
-      if (declarationFile.endsWith(JSON_EXTENSION)) {
-          return DeclarativeUtil.loadDeclaration(declarationFile, MappingUtil.JSON_MAPPER, parsedProperties);
-      } else {
-          // default is YAML
-          return DeclarativeUtil.loadDeclaration(declarationFile, MappingUtil.YAML_MAPPER, parsedProperties);
-      }
+        // parse declaration
+        if (declarationFile.endsWith(JSON_EXTENSION)) {
+            return DeclarativeUtil.loadDeclaration(declarationFile, MappingUtil.JSON_MAPPER, parsedProperties);
+        } else {
+            // default is YAML
+            return DeclarativeUtil.loadDeclaration(declarationFile, MappingUtil.YAML_MAPPER, parsedProperties);
+        }
     }
 
     public void setDeclarationFiles(List<Path> declarationFiles) {